oe-buglabs: added and updated scripts for easier automation setup.

+++ buglabs-scripts/oe-configure.sh
@@ -3,23 +3,24 @@
 # Script to install OpenEmbedded for BUG
 # usage: oe-setup.sh install_dir package_src_dir tmp_dir
 
-OE_DIR=$1
-OE_SOURCES=$2
-OE_TMP=$3
+OE_SOURCES=$1
+OE_TMP=$2
 
-if [ -d ${OE_DIR} ]
-then
-        echo "Installation directory already exists, refusing to continue."
-        #exit -1
+if [ ! -d oe-buglabs ]; then
+	echo "Not in build root directory."
+	exit 1
 fi
 
-mkdir -p ${OE_DIR}
-cd ${OE_DIR}
-git clone git://github.com/buglabs/oe-buglabs.git
-cd oe-buglabs/
-git checkout -b buglabs-sw2.1 origin/buglabs-sw2.1
-cd ..
-source oe-buglabs/buglabs-scripts/reinstate-build-env
+if [ -f conf/local.conf ]; then
+	echo "Configuration already exists, not changing."
+	exit
+fi
+
+if [ ! -d conf ]; then
+	mkdir conf
+fi
+
+cp oe-buglabs/buglabs-scripts/local.conf conf/
 
 # Modify the script variables as specified in parameters
 mv conf/local.conf conf/local.conf.tmp
@@ -31,4 +32,4 @@ echo "TMPDIR=\"$OE_TMP\"" >> conf/local.conf
 cat conf/local.conf.tmp >> conf/local.conf
 rm conf/local.conf.tmp
 
-echo "Bug Labs OE build system setup.  Proceed to bitbaking."
+echo "Bug Labs OE build system configuration complete."
